Add Python support to PluginLoader
All plugins must now inherit from PluginBase and that is enforced by the
PluginLoader. For now, the sole purpose of PluginBase is to provide a
const optional<string> field: type_query. If type_query is present, it
is used to perform the TypeQuery in order for the plugin loader to pass
the plugin into Python.

By default the type_query string is not present, and when this is the
case the plugin loader will return None when asked about that class from
Python regardless of its availability from C++.